backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
supabase>=2.4.0
xgboost>=2.0.3
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
rapidfuzz>=3.6.0
httpx>=0.27.0
